+++ uc-non-persistent/ansible/tests/monitor/run.monitor.ping.sh
@@ -1,8 +1,10 @@
-#!/bin/bash
+#!/usr/bin/env bash
 # ---------------------------------------------------------------------------------------------
 # MIT License
 # Copyright (c) 2020, Solace Corporation, Ricardo Gomez-Ulmke (ricardo.gomez-ulmke@solace.com)
 # ---------------------------------------------------------------------------------------------
+
+trap "" SIGKILL
 
 echo;
 echo "##############################################################################################################"
@@ -18,30 +20,28 @@ echo
     projectHome=${scriptDir%/ansible/*}
     sharedSetupDir="$projectHome/shared-setup"; [ ! -d $sharedSetupDir ] && (echo ">>> ERROR: directory $sharedSetupDir DOES NOT exists."; exit)
 
-    export ANSIBLE_LOG_PATH="./ansible.log"
+    if [ -z "$RUN_LOG_DIR" ]; then export RUN_LOG_DIR=$scriptDir/tmp; mkdir $RUN_LOG_DIR > /dev/null 2>&1; fi
+    export ANSIBLE_LOG_PATH="$RUN_LOG_DIR/$scriptName.ansible.log"
     export ANSIBLE_DEBUG=False
     export ANSIBLE_HOST_KEY_CHECKING=False
+
+############################################################################################################################
+# Check if monitor running
 
 ############################################################################################################################
 # Environment Variables
 
     if [ -z "$1" ]; then
       if [ -z "$UC_NON_PERSISTENT_INFRASTRUCTURE" ]; then
-          echo ">>> missing infrastructure info. pass either as env-var: UC_NON_PERSISTENT_INFRASTRUCTURE or as argument"
+          echo ">>> ERROR: missing infrastructure info. pass either as env-var: UC_NON_PERSISTENT_INFRASTRUCTURE or as argument"
           echo "    for example: $scriptName azure.infra1-standalone"
           echo; exit 1
       fi
     else
       export UC_NON_PERSISTENT_INFRASTRUCTURE=$1
     fi
-
-    if [ -z "$runId" ]; then
-      export runId=$(date -u +%Y-%m-%d-%H-%M-%S)
-    fi
-
-    if [ -z "$runStartTsEpochSecs" ]; then
-      export runStartTsEpochSecs=$(date -u +%s)
-    fi
+    if [ -z "$RUN_ID" ]; then export RUN_ID=$(date -u +"%Y-%m-%d-%H-%M-%S"); fi
+    if [ -z "$runStartTsEpochSecs" ]; then export runStartTsEpochSecs=$(date -u +%s); fi
 
 ##############################################################################################################################
 # Prepare
@@ -59,15 +59,16 @@ rm -f $resultDir/ping-stats.*.json
   privateKeyFile=$(assertFile "$projectHome/keys/"$cloudProvider"_key") || exit
 
   ansible-playbook \
+                  --fork 1 \
                   -i $inventoryFile \
                   --private-key $privateKeyFile \
                   $playbook \
                   --extra-vars "RESULT_DIR=$resultDir" \
-                  --extra-vars "RUN_ID=$runId" \
+                  --extra-vars "RUN_ID=$RUN_ID" \
                   --extra-vars "RUN_START_TS_EPOCH_SECS=$runStartTsEpochSecs" \
                   --extra-vars "INVENTORY_FILE=$inventoryFile"
 
-  if [[ $? != 0 ]]; then echo ">>> ERROR retrieving ping stats: $scriptName"; echo; exit 1; fi
+  code=$?; if [[ $code != 0 ]]; then echo ">>> ERROR - $code - playbook exit: $scriptName"; echo; exit 1; fi
 
   echo "##############################################################################################################"
   echo "# Results in: $resultDir"